ci: refactor resolve cs proj script to use env vars
